clean project creation helper functions into separate module inside deet/scripts/
