+++ deet/scripts/commands/project.py
@@ -1,7 +1,6 @@
 # ruff: noqa: PLC0415
 """Sub-commands for project initialisation and configuration."""
 
-import re
 from pathlib import Path
 from typing import TYPE_CHECKING, Annotated
 
@@ -11,170 +10,23 @@ if TYPE_CHECKING:
 
 import typer
 from InquirerPy import inquirer
-from pydantic import ValidationError
 
 from deet.processors.converter_register import SupportedImportFormat
-from deet.scripts.typer_context import project_required
-from deet.settings import DataExtractionSettings, LogLevel
-from deet.ui import fail_with_message, notify
-from deet.ui.terminal import (
-    console,
-    continue_after_key,
-    render_template,
-    run_model_wizard,
+from deet.scripts.project_utils import (
+    DataFormatOption,
+    DataPathOption,
+    ForceOption,
+    PdfDirOption,
+    create_project,
+    guard_overwrite,
+    prompt_name,
 )
-from deet.ui.terminal.components import info_panel, wizard_field_help
-from deet.ui.terminal.wizards import get_ui_metadata, inquire_pydantic_field
+from deet.scripts.typer_context import project_required
+from deet.settings import LogLevel
+from deet.ui import fail_with_message, notify
+from deet.utils.text import slugify
 
 app = typer.Typer(help="Commands to create and configure deet projects.")
-
-
-def slugify(value: str) -> str:
-    """Turn a project name into a filesystem-safe directory name."""
-    slug = re.sub(r"[^a-z0-9]+", "-", value.strip().lower()).strip("-")
-    if not slug:
-        invalid = f"Could not derive a directory name from {value!r}"
-        raise ValueError(invalid)
-    return slug
-
-
-def _run_init_wizard(root: Path, name: str) -> None:
-    """
-    Run the interactive project + credentials wizards and set the project up.
-
-    Prompts for every project field except ``name`` (supplied here), anchors the
-    project to ``root`` (re-expressing resource paths relative to it), then writes
-    the project structure and credentials into ``root``.
-    """
-    from deet.data_models.project import DeetProject
-
-    console.clear()
-    init_md = render_template("project/init")
-    console.print(info_panel(init_md, title=":speedboat: project set-up"))
-    continue_after_key()
-
-    project = run_model_wizard(DeetProject, prefill={"name": name})
-    project.anchor_to(root)
-    project.setup()
-
-    console.clear()
-    configure_env_md = render_template("project/configure_env.md")
-    console.print(info_panel(configure_env_md, ":key: Credential management"))
-    continue_after_key()
-    settings = run_model_wizard(DataExtractionSettings)
-    settings.dump_to_env(target_path=root / ".env")
-
-    new_directory = root.relative_to(Path.cwd())
-
-    console.clear()
-    console.print(
-        info_panel(
-            render_template(
-                "project/success.md", project=project, new_directory=str(new_directory)
-            )
-        )
-    )
-
-
-def _guard_overwrite(target_dir: Path, *, force: bool, interactive: bool) -> None:
-    """
-    Guard against overwriting an existing project at ``target_dir``.
-
-    Does nothing if ``force`` is set or no project exists there. Otherwise prompts
-    to overwrite when running interactively, or exits with guidance when headless
-    (where prompting is impossible).
-    """
-    from deet.data_models.project import PROJECT_FILE, DeetProject
-
-    if force or not (target_dir / PROJECT_FILE).exists():
-        return
-    if not interactive:
-        fail_with_message(
-            f"A project already exists at {target_dir}. Use --force to overwrite."
-        )
-    existing_project = DeetProject.load(target_dir)
-    notify(
-        (
-            f"Project {existing_project.name} already exists in {target_dir}. "
-            "Continuing could overwrite data and settings"
-        ),
-        level=LogLevel.WARNING,
-    )
-    if not inquirer.confirm("Overwrite existing project?").execute():
-        fail_with_message("Exiting..")
-
-
-def _create_project(
-    root: Path,
-    name: str,
-    *,
-    data_path: Path | None,
-    data_type: SupportedImportFormat,
-    pdf_dir: Path | None,
-) -> None:
-    """
-    Create and set a project named ``name`` up at ``root``.
-
-    When resource paths are supplied the project is built headlessly from them;
-    otherwise the interactive wizard collects the remaining fields. Either way the
-    project is anchored to ``root`` and its directory structure written there.
-    """
-    from deet.data_models.project import DeetProject
-
-    if any([data_path, pdf_dir]):
-        try:
-            if data_path is None:
-                fail_with_message(
-                    "Gold-standard data (--data) is required to create a project"
-                    " non-interactively"
-                )
-            project = DeetProject(
-                name=name,
-                gold_standard_data_path=data_path,
-                gold_standard_data_format=data_type,
-                pdf_dir=pdf_dir,
-            )
-        except ValidationError as e:
-            fail_with_message(f"Invalid project configuration:\n{e}")
-        project.anchor_to(root)
-        project.setup()
-    else:
-        _run_init_wizard(root, name)
-
-
-# Shared options, so `init` and `new` accept identical headless arguments.
-DataPathOption = Annotated[
-    Path | None,
-    typer.Option(
-        "--data",
-        "-d",
-        exists=True,
-        file_okay=True,
-        dir_okay=False,
-        resolve_path=True,
-        help="Path to your gold standard annotation data",
-    ),
-]
-DataFormatOption = Annotated[
-    SupportedImportFormat,
-    typer.Option("--format", "-t", help="Format of your gold standard annotated data."),
-]
-PdfDirOption = Annotated[
-    Path | None,
-    typer.Option(
-        "--pdfs",
-        "-p",
-        exists=True,
-        file_okay=False,
-        dir_okay=True,
-        resolve_path=True,
-        help="The folder where your pdfs for data extraction are stored.",
-    ),
-]
-ForceOption = Annotated[
-    bool,
-    typer.Option("--force", "-f", help="Overwrite existing project data."),
-]
 
 
 @app.command()
@@ -190,31 +42,15 @@ def init(
 
     The project name is taken from the current directory. Leave the data and pdf
     options empty to enter the interactive wizard.
+    Use deet project new to create a project in a new directory.
     """
     root = Path.cwd()
-    _guard_overwrite(
+    guard_overwrite(
         root, force=force_overwrite, interactive=not any([data_path, pdf_dir])
     )
-    _create_project(
+    create_project(
         root, root.name, data_path=data_path, data_type=data_type, pdf_dir=pdf_dir
     )
-
-
-def _prompt_name() -> str:
-    """Prompt for a project name."""
-    from deet.data_models.project import DeetProject
-
-    info = DeetProject.model_fields["name"]
-    ui = get_ui_metadata(info)
-    if ui is None:
-        no_ui = "No UI component for name"
-        raise ValueError(no_ui)
-    ui_help = ui.help + (
-        ". The name you enter will be standardised and used to create a directory"
-    )
-    console.clear()
-    console.print(wizard_field_help("name", ui_help))
-    return str(inquire_pydantic_field(DeetProject, "name", info, ui))
 
 
 @app.command()
@@ -231,22 +67,20 @@ def new(
     """
     Create a new project in its own directory, named after ``name``.
 
-    A directory named after ``name`` (slugified) is created in the current
-    directory and the project is set up inside it, leaving the current directory
-    untouched. Omit ``name`` to be prompted for it. Leave the data and pdf options
+    Leave the data and pdf options
     empty to enter the interactive wizard. Use `deet project init` to turn the
     current directory into a project.
     """
     if name is None:
-        name = _prompt_name()
+        name = prompt_name()
 
     target = Path.cwd() / slugify(name)
-    _guard_overwrite(
+    guard_overwrite(
         target, force=force_overwrite, interactive=not any([data_path, pdf_dir])
     )
     target.mkdir(parents=True, exist_ok=True)
     notify(f"Creating project in {target}", level=LogLevel.INFO)
-    _create_project(
+    create_project(
         target, name, data_path=data_path, data_type=data_type, pdf_dir=pdf_dir
     )
